requirements.txt
pymupdf
pillow
requests
pandas
tqdm
